update price square cols and dtos to decimal and fix wager display

+++ RSS-Services/DTOs/CurrentGameSummaryDTO.cs
@@ -5,7 +5,7 @@ namespace RSS_Services.DTOs
         public Guid GameId { get; set; }
         public string GameName { get; set; }
         public string GameType { get; set; }
-        public int PricePerSquare { get; set; }
+        public decimal PricePerSquare { get; set; }
         public int SquaresClaimed { get; set; }
         public bool IsHost { get; set; }
         public bool IsOpen { get; set; }
